Refactor tests to use qgis_plugin_tools
